refactor(Layer): add PointCloudLayer.
